Extract logic from ApplicationController into digestable dependency files
